+++ lib/core.ps1
@@ -225,8 +225,8 @@ function cache_path($app, $version, $url) { "$cachedir\$app#$version#$($url -rep
 
 # apps
 function sanitary_path($path) { return [regex]::replace($path, "[/\\?:*<>|]", "") }
-function installed($app, $global) {
-    if (-not $PSBoundParameters.ContainsKey('global')) {
+function installed($app, [Nullable[bool]]$global) {
+    if ($null -eq $global) {
         return (installed $app $false) -or (installed $app $true)
     }
     # Dependencies of the format "bucket/dependency" install in a directory of form
@@ -507,12 +507,12 @@ function Invoke-ExternalCommand {
     }
     $Process = New-Object System.Diagnostics.Process
     $Process.StartInfo.FileName = $FilePath
-    $Process.StartInfo.Arguments = ($ArgumentList | Select-Object -Unique) -join ' '
     $Process.StartInfo.UseShellExecute = $false
     if ($LogPath) {
-        if ($FilePath -match '(^|\W)msiexec($|\W)') {
-            $Process.StartInfo.Arguments += " /lwe `"$LogPath`""
+        if ($FilePath -match '^msiexec(.exe)?$') {
+            $ArgumentList += "/lwe `"$LogPath`""
         } else {
+            $redirectToLogFile = $true
             $Process.StartInfo.RedirectStandardOutput = $true
             $Process.StartInfo.RedirectStandardError = $true
         }
@@ -520,9 +520,32 @@ function Invoke-ExternalCommand {
     if ($RunAs) {
         $Process.StartInfo.UseShellExecute = $true
         $Process.StartInfo.Verb = 'RunAs'
+    } else {
+        $Process.StartInfo.CreateNoWindow = $true
+    }
+    if ($FilePath -match '^((cmd|cscript|wscript|msiexec)(\.exe)?|.*\.(bat|cmd|js|vbs|wsf))$') {
+        $Process.StartInfo.Arguments = $ArgumentList -join ' '
+    } elseif ($Process.StartInfo.ArgumentList.Add) {
+        # ArgumentList is supported in PowerShell 6.1 and later (built on .NET Core 2.1+)
+        # ref-1: https://docs.microsoft.com/en-us/dotnet/api/system.diagnostics.processstartinfo.argumentlist?view=net-6.0
+        # ref-2: https://docs.microsoft.com/en-us/powershell/scripting/whats-new/differences-from-windows-powershell?view=powershell-7.2#net-framework-vs-net-core
+        $ArgumentList | ForEach-Object { $Process.StartInfo.ArgumentList.Add($_) }
+    } else {
+        # escape arguments manually in lower versions, refer to https://docs.microsoft.com/en-us/previous-versions/17w5ykft(v=vs.85)
+        $escapedArgs = $ArgumentList | ForEach-Object {
+            # escape N consecutive backslash(es), which are followed by a double quote, to 2N consecutive ones
+            $s = $_ -replace '(\\+)"', '$1$1"'
+            # escape N consecutive backslash(es), which are at the end of the string, to 2N consecutive ones
+            $s = $s -replace '(\\+)$', '$1$1'
+            # escape double quotes
+            $s = $s -replace '"', '\"'
+            # quote the argument
+            "`"$s`""
+        }
+        $Process.StartInfo.Arguments = $escapedArgs -join ' '
     }
     try {
-        $Process.Start() | Out-Null
+        [void]$Process.Start()
     } catch {
         if ($Activity) {
             Write-Host "error." -ForegroundColor DarkRed
@@ -530,11 +553,17 @@ function Invoke-ExternalCommand {
         error $_.Exception.Message
         return $false
     }
-    if ($LogPath -and ($FilePath -notmatch '(^|\W)msiexec($|\W)')) {
-        Out-UTF8File -FilePath $LogPath -Append -InputObject $Process.StandardOutput.ReadToEnd()
-        Out-UTF8File -FilePath $LogPath -Append -InputObject $Process.StandardError.ReadToEnd()
+    if ($redirectToLogFile) {
+        # we do this to remove a deadlock potential
+        # ref: https://docs.microsoft.com/en-us/dotnet/api/system.diagnostics.process.standardoutput?view=netframework-4.5#remarks
+        $stdoutTask = $Process.StandardOutput.ReadToEndAsync()
+        $stderrTask = $Process.StandardError.ReadToEndAsync()
     }
     $Process.WaitForExit()
+    if ($redirectToLogFile) {
+        Out-UTF8File -FilePath $LogPath -Append -InputObject $stdoutTask.Result
+        Out-UTF8File -FilePath $LogPath -Append -InputObject $stderrTask.Result
+    }
     if ($Process.ExitCode -ne 0) {
         if ($ContinueExitCodes -and ($ContinueExitCodes.ContainsKey($Process.ExitCode))) {
             if ($Activity) {
@@ -604,12 +633,12 @@ function movedir($from, $to) {
     $proc.StartInfo.RedirectStandardError = $true
     $proc.StartInfo.UseShellExecute = $false
     $proc.StartInfo.WindowStyle = [System.Diagnostics.ProcessWindowStyle]::Hidden
-    $proc.Start()
-    $out = $proc.StandardOutput.ReadToEnd()
+    [void]$proc.Start()
+    $stdoutTask = $proc.StandardOutput.ReadToEndAsync()
     $proc.WaitForExit()
 
     if($proc.ExitCode -ge 8) {
-        debug $out
+        debug $stdoutTask.Result
         throw "Could not find '$(fname $from)'! (error $($proc.ExitCode))"
     }