+++ libexec/scoop-status.ps1
@@ -1,5 +1,8 @@
 # Usage: scoop status
 # Summary: Show status and check for new app versions
+# Help: Options:
+#   -l, --local         Checks the status for only the locally installed apps,
+#                       and disables remote fetching/checking for Scoop and buckets
 
 . "$PSScriptRoot\..\lib\manifest.ps1" # 'manifest' 'parse_json' "install_info"
 . "$PSScriptRoot\..\lib\versions.ps1" # 'Select-CurrentVersion'
@@ -9,6 +12,8 @@ $currentdir = fullpath $(versiondir 'scoop' 'current')
 $needs_update = $false
 $bucket_needs_update = $false
 $script:network_failure = $false
+$no_remotes = $args[0] -eq '-l' -or $args[0] -eq '--local'
+if (!(Get-Command git -ErrorAction SilentlyContinue)) { $no_remotes = $true }
 $list = @()
 if (!(Get-FormatData ScoopStatus)) {
     Update-FormatData "$PSScriptRoot\..\supporting\formats\ScoopTypes.Format.ps1xml"
@@ -27,10 +32,13 @@ function Test-UpdateStatus($repopath) {
     }
 }
 
-$needs_update = Test-UpdateStatus $currentdir
-foreach ($bucket in Get-LocalBucket) {
-    if (Test-UpdateStatus (Find-BucketDirectory $bucket -Root)) {
-        $bucket_needs_update = $true
+if (!$no_remotes) {
+    $needs_update = Test-UpdateStatus $currentdir
+    foreach ($bucket in Get-LocalBucket) {
+        if (Test-UpdateStatus (Find-BucketDirectory $bucket -Root)) {
+            $bucket_needs_update = $true
+            break
+        }
     }
 }
 
@@ -38,7 +46,7 @@ if ($needs_update) {
     warn "Scoop out of date. Run 'scoop update' to get the latest changes."
 } elseif ($bucket_needs_update) {
     warn "Scoop bucket(s) out of date. Run 'scoop update' to get the latest changes."
-} elseif (!$script:network_failure) {
+} elseif (!$script:network_failure -and !$no_remotes) {
     success 'Scoop is up to date.'
 }
 
@@ -57,7 +65,7 @@ $true, $false | ForEach-Object { # local and global apps
         $item.'Installed Version' = $status.version
         $item.'Latest Version' = if ($status.outdated) { $status.latest_version } else { "" }
         $item.'Missing Dependencies' = $status.missing_deps -Split ' ' -Join ' | '
-        $info = $()
+        $info = @()
         if ($status.failed)  { $info += 'Install failed' }
         if ($status.hold)    { $info += 'Held package' }
         if ($status.removed) { $info += 'Manifest removed' }